use of static cast

+++ src/lexer.cpp
@@ -30,10 +30,10 @@ template <typename Lexer> Lex::ParseInput<Lexer>::ParseInput() {
 
 template <typename Token>
 bool Lex::Add::operator()(Token const& t, std::vector<std::string>& tks) const {
-    if (t.id() == symbol_table::i_) {
+    if (static_cast<unsigned long>(t.id()) == symbol_table::i_) {
         return true;
     }
-    tks.push_back(symbol_table::token_types_r_.at(t.id()));
+    tks.push_back(symbol_table::token_types_r_.at(static_cast<unsigned long>(t.id())));
     return true;
 }
